opdate

+++ catenary/catenary.go
@@ -6,5 +6,5 @@ type Catenary struct {
 	Profile      profile.Profile `json:"profile"`
 	Slack        float64         `json:"slack"`
 	MaxSag       float64         `json:"max_sag"`
-	Tessellation float64         `json:"tessellation"`
+	Tessellation float32         `json:"tessellation"`
 }
